Use `ensure_packages()` to avoid resource conflicts
This module uses an older technique to make sure that `curl` and
`mercurial` are installed; unfortunately, this technique causes a
resource conflict in the event that either of those package resources
are declared elsewhere.  This patch declares a dependency on
`puppetlabs/stdlib` and then uses the `ensure_packages()` function to
achieve the same effect without the resource conflict.